pyproject.toml
[project]
name = "reasoning-over-slaterecs"
version = "0.1.0"
description = ""
authors = [
    {name = "Asr419",email = "asrthebest17@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "recsys-slates-dataset (>=1.0.3,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyarrow (>=19.0.0,<20.0.0)",
    "openai (>=1.60.2,<2.0.0)",
    "scipy (>=1.7,<2.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "gensim (==4.3.3)",
    "torch[cu121] (>=2.2.0,<3.0.0)",
    "transformers (>=4.30.0,<5.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "fasttext (>=0.9.3,<0.10.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
